requirements.txt
langchain>=0.1.0
langchain-chroma>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.1
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
pypdf>=3.17.0
python-dotenv>=1.0.0
chromadb>=0.4.22
sentence-transformers>=2.2.0
